Separate get-events and get-last-event functions
By re-using the same function, we ended up in a situation where we
paginated through all events simply to return the id of the first one of
the first page we received.

This was leading to some Throttling errors when working out the first
last-event-id.

With dedicated functions, we can simplify both paths and avoid the
inefficiency.