Support Numeric parameters in StackSpecYaml
An actual CloudFormation template can accept Strings or Numbers (the
other types will be handled as Strings when Yaml).

https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/parameters-section-structure.html#parameters-section-structure-properties-type

We need an Amazonka Parameter, which is always Text. And our direct
parser meant users could not put bare numbers in a StackSpecYaml, even
if that's what the template defines. The workaround was easy, to quote
the value no matter what, but that could be confusing and we could be a
little nicer through our parser.