Use Blammo's Env.parseWith for the stderr default
The way it was before, we always logged to `stderr` no matter what. That
was because Stackctl produces useful `stdout` (e.g. `changes`), and we
don't want to include logging when something captures it. This behaves
the same but will respect `LOG_DESTIONATION` if set.

This matters over in Platform CLI where we are performing integration
tests on the CLI's output. When logging and errors are combined we get
confusing failures (like `jq` parse failures) when there are other
errors (like the test AWS SSO session expired). Being able to put this
logging to `stdout` there will solve that.