Add PATH argument to changes sub-command
This is the only case of "useful" output we produce. Making its
destination explicit will free us up to simplify other outputs (logging
and stack-deployment events) to hopefully address the ordering issues we
see by the careful use of handles we needed to support the default
output for changes.

In our own use of the `changes` subcommand, we redirect it to a file
anyway, so this will even make that more convenient.