Tweak output in runAction errors
Throwing an exception, while simple, does not use `displayException`
(bug?), so we explicitly logged `LambdaInvokeFailure` so we could
pretty-print it and then threw something as a simple exit mechanism.

The output can be confusing since you get pretty logging, then a dense
garble of text on the exit. We can just throw a bare constructor and
achieve the same with less confusing output. Doing that for
`NoSuchOutput` too makes things consistent and affords pretty-printed
logging of that case as well.