Add --parameter to changes and deploy
We originally did not have this because the expectation is that you are
always reviewing, committing, and automatically-executing things from
the on-disk representation.

However, RDS upgrades (as one example) require a few manual deployments
with different parameters. Being able to pass an override for one of the
deploys is better than having to deploy-edit-deploy-edit-deploy,
particularly when you're trying to script the whole cycle.