Add --no-flip to stackctl-capture
Sometimes we are capturing legacy templates with JSON bodies in an
attempt to move from old to new (Stackctl-based) tooling. If we flip
these to Yaml, CloudFormation often infers spurious changes that makes
it riskier to proceed with the capture. Keeping them JSON avoids this
and lets us tackle flipping to Yaml later while still getting over to
the new tooling now.

Without this, I've been using `capture` then manually copying in the
JSON and renaming the file. This is annoying and error-prone.

I debating calling the flag `--no-flip-yaml`, as in "Don't flip to
Yaml", but then I thought `--no-flip-json`, as in "Don't flip from JSON"
could be equally reasonable. So I went with being _more_ vague in hopes
the user will check the docs if they're unsure.