Support Description in StackSpecs
Descriptions are useful to attach to Stacks, we should support that in
our specs. If users put a `Description` in their `templates/`, we have
(and will continue to) respect those simply by consequence of
CloudFormation doing so, but I think those ergonomics are wrong.

If we think of stacks as instantiations of re-used templates, the fact
that the description lives as a property on the template instead of the
stack is almost certainly a historical accident. What good is describing
all of your auto-scaling groups, which make use of a nice parameterized
template, as "An Auto-scaling group"?

Stackctl affords us the ability to correct for that. We accept
`Description` in the stack-specific specs, and (so long as there's no
`Description` already in the template being used), we write that in on
deployment of that template.

Inserting the descriptions in Yaml is done carefully. We do a naive
textual insert of a `Description` key at the start of the Yaml, rather
than the more robust parse-update-render. The reason is we want to
preserve any formatting and comments from the on-disk representation in
the in-cloud version. This feels worthwhile. Besides preventing surprise
and confusion for anyone looking at the in-cloud representation, it also
ensures that creating change sets when there are no changes will remain
fast. I have no proof, but AWS _seems_ to leverage some kind of template
checksum to make the no-changes case faster.