Strengthen ParameterYaml type
Because `ToJSON` on `ParameterYaml` had to deal with a potentially null
key in the underlying `Parameter`, we used `object []` for the case of
either side being missing. Having a Key but not a Value is actually a
realistic use-case, and we were generating:

```yaml
Parameters:
 - {}
```

When we should've generated:

```yaml
Parameters:
  - ParameterKey: Something
    ParameterValue: null
```

Or:

```yaml
Parameters:
  - ParameterKey: Something
```

To make this easier to reason about, we upgraded `ParameterYaml` to
`data` so we could hold a non-null key and possibly-null value. We wrote
`unParameterYaml` to behave the same and a `parameterYaml` to fail _on
construction_ for the invalid case of a missing key.

This makes it more natural to handle a missing value correctly.