Move filtering later, operate on StackSpec
This will allow us to treat filter arguments in other ways, such as
matching Stack names or template paths too. Matching names is more
intuitive for users, and matching template paths means we can remove the
logic of finding changed stacks when template paths change on CI; we can
simply pass the template paths to `--filter` and it'll do the Right
Thing.