Move "Last" wrapper up to ParameterYaml{pyValue}
A `ParmeterValue` shouldn't care how its combined, its when the
key-collision of two `ParmaterYaml` values happens that it matters; so
this makes more sense there.

I only didn't have it there originally because I was finding a test
failure in parsing `ParameterValue`, but I had something else broken, it
seems.