Fix JSON instance for StackSpecYaml
We didn't define a custom `ToJSON` for the `ParamtersYaml` or `TagsYaml`
objects to generate the more natural key-value style, but we _did_
define customer instances on the `ParameterYaml` and `TagYaml` element
types. This resulted in some weird JSON during generation or capture.
This commit fixes that, relying on the new failing test to be sure.