Fix JSON bugs in Spec generation
* Fix invalid FromJSON(Action)

We were loading the same constructor by either tag. I also moved from
asum to (<|>), which produces better failures

* Fix JSON instance for StackSpecYaml

We didn't define a custom `ToJSON` for the `ParamtersYaml` or `TagsYaml`
objects to generate the more natural key-value style, but we _did_
define custom instances on the `ParameterYaml` and `TagYaml` element
types. This resulted in some weird JSON during generation or capture.

* Update writeStackSpec to use ssSpecRoot

It has the same values as the separate `parent` argument, which predated
it. Doing this means that the `stackSpec...File` functions are only used
in contexts where them being relative makes more more sense.

And being relative is also required to fix `--filter`.

* Add spec on filtering from the output of generate

This failed until the previous commit.

* Fix template use-case that still needs the root

* Add some logging around ChangeSet creation

* Avoid more direct accesses of StackSpec record fields