Push runAppT into Subcommand{run}
When all `Subcommand`s are called with `runAppT`, it meant we always had
to initialize the `App` no matter the subcommand. This led to a
surprising scenario where even something trivial like `stack version`
requires a valid AWS connection. Oops.

Making `Subcommand{run}` call `runAppT` itself is no more complicated
(especially with the `runAppSubcommand` helper), in fact it makes the
constraints in the `Commands` module simpler, and it naturally allows
for some subcommands to _not_ do that (like `version`) and so not incur
the AWS requirements.