Add --no-include-full option to stackctl-changes
The purpose of this command is primarily to build a PR comment to post
to PRs that change resources. If the comment is no different than the
last one (e.g. you make a fixup in an existing PR), our add-pr-comment
action could easily skip adding a duplicate comment. Problem is, the
full JSON contains identifiers that are different every time, causing
the comment body to never be an actual duplicate.

I've never found these details particularly valuable, so let's have an
option to omit them, making it a little easier for the duplicate-comment
logic to work for us.