Delete corresponding Stacks when specs are removed
Given:

- A full path to `stacks/.../foo/bar.yaml` is included in `--filter`
- And that path does not exist
- But the corresponding, conventionally-named Stack `foo-bar` does

Then `stackctl` will handle this as a delete, meaning it'll be shown as
such in `changes` and processed in `deploy`.

This is difficult to trigger as a manual operator, which is fine. The
idea is that operators will remove files and open a PR. Since the
removed file will be in the PR changed-files, and so included in
`--filter`, the above will apply and required removals will be processed
like any other gitops change.