Replace pandoc-based man-page processing with ronn
Ronn is a markdown-like format for authoring man-pages. The main benefit
to this approach is its nicely-styled, cross-linked html output format,
which I'm deployed as a GH Pages site for the project now. The main
downside is a slight decrease in quality of installed man-pages, which
we consider acceptable.

NOTE: The original `ronn` gem is no longer maintained, but development
continues in a new `ronn-ng` fork, which we're using here.