Move to a different release action
The way things were before we created a (pulished) release first, then
build and uploaded assets. This meant (particularly with OSX) the
release existed as "latest" for a long time without assets. This causes
issues for anyone that tries to install during this time, including some
test suites.

This new action allows you to create a Draft release, upload assets,
then publish it -- meaning it won't be visible until it's ready, solving
that problem.

Annoyingly, we had to fork the action to do it, as the best one our
there needed a fix and the maintainer is gone.