Handle AWS ServiceError to more readable exit
Any un-handled AWS errors would crash with a shown exception that is
difficult to read. This handler just reformats it to a more readable
error-log before still exiting failure.

We implemented this in a downstream user of the library, but it makes
sense to be in use all the time and provided by us.