Add golden specs on formatChangeSet
Use [hspec-golden][] to assert formatting of example changesets (so far
just one) doesn't change unexpectedly. This will allow safe refactoring
and extension of this content without having to generate actual
changesets and/or GitHub comments to test.

Approach: every `.json` file in the given directory is formatted and the
result compared against the equivalent `.txt` file (for `--format=tty`)
or a `.md` file (for `--format=pr`).

[hspec-golden]: https://hackage.haskell.org/package/hspec-golden-0.2.1.0