Add awsWithAuth
This function allows access to the underlying `AuthEnv`. It's not used
by Stackctl (yet) but is used by tooling that uses Stackctl as its AWS
SDK wrapper to supply concrete keys to external processes that aren't
SSO-session aware.

In the future, this AWS "library" should be extracted and shared, but
for now we accept such minor extensions even if they only exist for
these external. use-cases.