Simplify awsSimple
Amazonka-2.0 adds a `Typeable` constraint to its `send`-et-al functions,
presumably because it's using `typeRep` to render the type of response
in its own logging. This means we too can use `typeRep` for our own
error message in `awsSimple` and thus not need to pass in an argument
for that.

Passing this argument was always an annoying bit of boilerplate and
subject to inconsistencies (sometimes we'd prefix with the service,
sometimes not, for example). This is more convenient and consistent. The
only downsides are churn for callers and requiring a major version bump,
which is acceptable.