Fix auto-expansion of filter option
Using (e.g.) `--filter progress` is meant to automatically find all
stacks related to this by turning that into a series of globs. Part of
that was turning it into `progress/**`, which is meant to find all
sub-directories. This was incorrect and needed to be `progress/**/*` for
that to work reliably. This small commit fixes that.