Add OneOrListOf, use for Action{run}
Sometimes it's useful to do more than one thing on an event, such as
create and migrate the DB on post-deploy of a DB-related resource.

Specifying a list for `run` is simpler, with more obvious semantics,
than specifying multiple actions with the same `on`. However, to avoid
breaking existing specifications specifying a single item for `run`
should still be supported.

The `OneOrListOf` wrapper implements parsing such values either way
either. And it's `Foldable` instance makes usage very nice. The
`Semigroup` instance is not currently used, so I'm open to removing it,
but it seemed reasonable to include.