Handle no-or-missing specs more safely
While filtering down to no specs is expected (and handled), having no
specs at all is not expected. And we now use this pre-filtered list with
`minimumBy` in `checkForUnknownDepends`. So if we found truly no specs
at all, we would have an exception there trying to get the minimum of an
empty list.

I've made this mistake myself, *thinking* I was in our `infra`
directory, running `stackctl` repeatedly with different forms of
`--filter` that I thought were not doing what I expected, only to
realize I'm in the wrong directory.

Without this change, that would now be an even more confusing empty-list
exception from the introduction of `minimumBy`. With this change, it's
back to the no-op it was before, but now has an even clearer warning,
which was added as an obvious consequence of thinking through the types
at all and arriving at a `Nothing` that had to be handled.