Respect STACKCTL_DIRECTORY in findRemovedStack
When inferring removed stacks, we use use the fact that a specification
doesn't exist on disk to decide to remove the corresponding stack.
Checking for the specification on disk was not using
`STACKCTL_DIRECTORY`, so if one were set this would always report the
file as missing and could result in deleting a stack we should not.

Ideally, I would have added a failing test before fixing this, but this
area is just not easy to get a test on at this time.