chore: add debug logging before AWS.discover
There is a bug[^1] in the development version of Amazonka, which we have
to use to get latest GHC support, that causes `newEnv` to perform a slow
DNS lookup, attempting to see if it's being run on an EC2 instance.

Since this happens early, before any logging, it can appear our CLIs are
stuck for about 30s without any output. And bumping `LOG_LEVEL` doesn't
help. Adding a message here will at least provide *some* feedback, while
we seek to address this for real.

[^1]: https://github.com/brendanhay/amazonka/issues/1018